% Generated by build script
\name{gt_committees}
\alias{gt_committees}
\title{List congressional committees}
\usage{
gt_committees()
}
\arguments{
\item{committee_type}{"senate", "house", or "joint" (default: all)}
\item{obsolete}{Include obsolete committees? (default FALSE)}
\item{limit}{Results per page (default 100)}
}
\value{tibble: code, name, committee_type, url, obsolete}
\description{List congressional committees}
