% Generated by build script
\name{gt_votes}
\alias{gt_votes}
\title{Search congressional votes}
\usage{
gt_votes()
}
\arguments{
\item{congress}{Congress number}
\item{chamber}{"house" or "senate"}
\item{limit}{Results per page (default 50)}
\item{offset}{Pagination offset (default 0)}
\item{order}{Sort field (default "-created" for newest first)}
}
\value{tibble: id, congress, session, chamber, number, question,}
\description{Search congressional votes}
